Add files via upload
Knight.cs which inherits from Player.cs, is the settings for the Knight character in a WIP project of mine. The "knight" places blocks and to be able to do that, this script places a transparent visualization of said block which is constantly updated as the character moves. Using raytracing for this behavior was a great deal of fun and is why part of this script is working with masks to properly place said blocks